bugs fixed

See merge request ticket-managment/ticket-managment!251

+++ core/src/main/java/com/sipl/ticket/core/dao/repository/WorkflowInstanceRepository.java
@@ -28,7 +28,7 @@ public interface WorkflowInstanceRepository extends JpaRepository<WorkflowInstan
             @Param("entityType") String entityType,
             @Param("status") Integer status,
             @Param("userId") Long userId,
-            @Param("roleId") Long roleId,
+            @Param("roleId") Integer roleId,
             Pageable pageable
     );
